POC for single source of truth for documentation
With the current docs for Barman 3.11 we are facing an issue: we use
different files to build the HTML manual and the Linux man pages. As
a consequence, we need to keep updating the same information across
diffrent sources for different docs.

This commit performs a POC for having a single source of truth when
building the Barman documentation. We use reStructuredText markup
language along with Sphinx to build any kind of documentation, i.e.:

* HTML docs (complete)
* PDF docs (do not include code API docs)
* Linux man pages (include only command and configuration reference)

The `index.rst` file defines the complete structure of the docs. Then,
through the variable `exclude_from_builder` in `conf.py` we are able
to define a blacklist of files for each different builder.

To build the documentation, we only need to run either of the
following commands, depending on the desired target:

* `tox -e docs -- html`: for HTML
* `tox -e docs -- latexpdf`: for PDF
* `tox -e docs --  man`: for Linux man pages

References: BAR-299.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>