Add commands to documentation
This commit adds all commands available in barman,
the description and the details on options and fields
available for each command.

The commands section (commands.rst) starts with a
synopsis on the `barman` command, followed by the
global options, a shortcut table and a brief on exit
statuses. Lastly, there is a list of the sub-commands
with synopsis, description and parameters.

Each command has its own `.rst` file which use the
`.. include::` directive from sphinx to be able to
group the commands in the `commands.rst` file. This
was done to be able to produce single man pages for
each command.

In the `.conf` file, the commands are referenced to
tell sphinx to render man pages.

References: BAR-311

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>