Use isoformat when producing json output in cloud-show-backup
This patch modifies the `to_json()` method of the BackupInfo class
substituting the `ctime` method from the datetime class with the
`isoformat` method, when printing `begin_time` and `end_time`
in json format.

The new format preserves all the relevant informations of the datetime
object, including the timezone.

References: bar-316

Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>