Add barman-cloud documentation
This commit adds documentation about the barman-cloud.

It was divided as follows:

* The main section is called `Barman for the cloud`. There is
an introduction on how Barman can work with cloud backups.
* Next section is called `barman-cli-cloud` that explains about
the package, followed by an `Installation` section.
* Lastly, the `Commands Reference` details each command
available from the package.

Each command was created in its own `.rst` file and they are
referenced in the `conf.py` file. This is meant to be used
to create MAN pages for each command.

This commit also adds the `--no-partial` to documentation
with the parameters description and alsoa brief explanation
in about its use.

References: BAR-312 & BAR-180

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>