Fix issues in Flake8 configuration for `super-linter`
This issue was identified while working on BAR-289. The `linter.yml`
is using a `.flake8` file which contains a set of configuration "A",
while `tox` is using a `tox.ini` file which contains another set of
configuration "B".

This commit fixes that oversight. Now both `super-linter` and `tox`
will use the very same `tox.ini`, so we have a single source of
truth for the Flake8 configuration.

References: BAR-315.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>