Create a stub for the "Contributing to Barman" section
We used to have a link in the main toctree called "Barman API docs",
which pointed to the content generated by `autodoc` extension based
on the barman docstrings.

Through this commit we replace that toctree entry with an entry for
a new section called "Contributing to Barman". This sections intends
to provide complete information about how a developer can contribute
to the Barman project.

The old "Barman API docs" was moved under this new section.

References: BAR-320.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>