Add a glossary to the documentation
This commit leverages from the Sphinx capability of maintaining a
glossary.

We add a new `glossary.rst` file where the glossary terms are
maintained. For sake of example, we add a couple of terms in the
glossary (Barman and PGDG).

This is the basic workflow for using the glossary:

* To add a new term to the glossary, add a new entry to the
  `glossary.rst` file.
* To reference a term from the glossary, just use the syntax
  :term:`NAME_OF_TERM`, and Sphinx creates a link to the Glossary
  for you.

References: BAR-326.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>