Explain the reasoning behind Check failure for WAL archives
This commit is adding the a reasoning note about why
we get `Check 'empty incoming directory' failed for server '<SERVER>'`
when WALs are present in the barman archive directory.

It also is formatting the documentation because of max sentence length,
some issues with the line breaks between bullet points and other house
keeping task so the actualy rendered doc looks nice and clean.

Closes GH-622 which exists in the barman repository (not this internal
one).

References: BAR-238

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>
Co-authored-by: Martín Marqués <martin.marques@enterprisedb.com>