Standardize references for the commands
This commit goal is to change the cross-reference labels to
the recommended standard detailed in the contributing
section.

The existing labels were all replaced and some rewording was
needed to accommodate the changes.

There are a few labels that have no topic written yet so it
has a `-TBD` appended to it provisionally.

References: BAR-379

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>