Refactoring barman-cli and barman-cli-cloud installation
This commit refactors the installation docs to accommodate
the barman-cli and barman-cli-cloud packages. there are
minor differences on how to install those packages depending
on the distribution used, so this is being explained in the
installation.rst file.

The barman_cloud installation section was also refactored to
reference the installation main section.

References: BAR-385

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>