Replace Material for Sphinx theme with PyData theme
Although the Material for Sphinx theme looked very nice, we had
an issue with it in regards to the TOC tree: either it renders
all tree nodes expanded, or it renders only the current section
nodes expanded.

We have been looking for alternative themes that could provide the
capability of expanding and contracting the tree nodes, and we found
these two:

* Read The Docs
* PyData

This commit delivers that feature by using PyData theme. We chose it
because the docs looked nicer on PyData than on Read The Docs theme.

It is worth noting that PyData renders the HTML docs a bit different
than Material and Read The Docs. Instead of merging the documents tree
with the page contents to create the left sidebar tree, PyData splits
the navigation through different parts:

* Top navbar: level-0 documents of the TOC tree.
* Left sidebar: level-1+ documents of the TOC tree.
* Right sidebar: active page content tree.

With that in mind, we had to apply some changes to the structure of
the docs:

* We had a single level-0 document defined in the root `index.rst`.
  Now we have these level-0 documents:
  * `user_guide/index.rst`: contains most of the previously existing
    documents, now as subdocuments under `user_guide/`.
  * `contributor_guide/index.rst`: the former `contributing.rst`, with
    a TOC tree. The subsections of the former `contributing.rst` were
    turned into separate subdocuments under `contributor_guide`. This
    was done because we had different logical pieces there, making sense
    to have them in separate subdocuments.
  * `release/index.rst`: the former `releases.rst`.
  * `faq/index.rst`: the former `faq.rst`.
  * `license/index.rst`: the former `license.rst`.

Besides that, these changes were applied to accomplish that goal:

* A new `index_pdf.rst` was created just so we have a different depth
  of the TOC tree between HTML and PDF outputs. As the `index.rst` is
  used by the main page of PyData, in that page we want only level-0
  heads. On the other hand, when building the index for the PDF, we want
  more levels for easy links to subsections.
* Image paths were fixed. They were using relative paths which no longer
  work, and that was preventing the PDF document to be built. Now we use
  absolute path when referencing images in the documentation.
* In Sphinx configuration:
  * Theme `sphinx_material` replaced with `pydata_sphinx_theme`
    * `html_theme_options` value changed to use options exposed by the
      new theme, in short:
      * Set up a couple icon buttons on the top nav.
      * Show 4 levels in the page content tree, and a edit page button
        in the right sidebar.
      * Hide the footer.
    * `html_side_bars`: to use the beauty left sidebar of PyData, with
      expandable/collapsable nodes.
  * `man_pages` now reference the new location of subdocuments, under
    the `user_guide` folder.
  * Autodoc of modules were moved from root to the `contributor_guide`
    folder.
    * Latex builder excludes was updated to reference the new path.
  * `latex_documents` was changed to reference the new `index_pdf.rst`
    file, which contains a different TOC tree depth than the default
    one used by HTML.

The `tox.ini` config was updated to reflect the usage of the new theme.

A side fix implemented by this commit: we fix the `exclude_patterns`
conf variable, so we don't exclude the `*.inc.rst` files when building
man pages. That is required so we are able to build both a complete man
page, and the split man pages.

References: BAR-394.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>