Change extension of included documentation files
When including files which contain reference anchors, sphinx will
complain that there are duplicate labels.

The workaround is to exclude those patterns. But in order to not
create a huge list of names in `exclude_patters` list, we changed
the extension of those included files from `.rst` to `inc..rst`,
and then just added `**/*.inc.rst` to the `exclude_patterns` list.

References: BAR-395

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>