Avoid s3 multipart uploads for small files
As it's been reported in the issue #1018, it seems that using the
`upload_fileobj` method to upload WAL files has more latency than using
`put_object`.

Various tests were run using a Python script similar to the one given in
the issue and the results were always similar to this:

The average latency for a file of 16 MB (default size of WAL files):
- using `upload_fileobj`: 4805.58967328ms
- using `put_object`: 3744.6347383500006ms

In this case, `put_object` is about 21% faster.

The average latency for file of 1GB (maximum size of relational files, by default):
- using `upload_fileobj`: 102020.6359049ms
- using `put_object`: 125556.6254884ms

In this case, `upload_fileobj` is about 18% faster.

So it really depends on the file size. I don't quite know the reason for
sure, but `upload_fileobj` uses multipart upload depending on the file
size while `put_object` always upload files in a single PUT request, so
that could add to the reason as multipart uploading requires some extra
steps.

AWS recommends considering multipart upload for files above 100MB. In
some testings I found out that the difference between both methods
really starts appearing for files around 150MB. But the results can vary
a lot from test to test, so I suppose it's better to rely on AWS numbers
as they probably have done better and more controlled tests than me in
order to write down that number.

That said, this commit slightly changes the upload method we call on the
boto3 library. We still use `upload_fileobj`, but now we pass a config
object containing a `multipart_threshold` value of 100MB, which
determines when multipart uploads should be triggered. When a file
smaller than that is uploaded, boto3 will actually trigger the
`put_object`  behind the scenes. Therefore, this is a simpler change as replacing
it for `put_object` directly would involve more code changes and the
methods don't have quite the same signature.

I also tested using `upload_fileobj` with different threshold values and
the results were pretty similar to the previous ones i.e. small files
being uploaded faster when multipart is not used and vice-versa, so it
must really use `put_object` behind the scenes.

Lastly, I also compared the changed code with the previous running on a
real Postgres cluster on an AWS instance, where network is more stable,
and the results were similar, showing a performance gain of around 15%
when avoiding multipart upload for 16MB WAL files.

Links:
- original issue: https://github.com/EnterpriseDB/barman/issues/1018
- AWS recommendation of when using multipart uploads: https://docs.aws.amazon.com/AmazonS3/latest/userguide/mpuoverview.html
- boto3 code which I think is used by `upload_fileobj`: https://github.com/boto/s3transfer/blob/develop/s3transfer/__init__.py#L727-L755
- `upload_fileobj` command ref: https://boto3.amazonaws.com/v1/documentation/api/latest/reference/services/s3/client/upload_fileobj.html

References: BAR-374

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>