Update man page generation when building packages
This commit:

1. Modify .gitignore to allow build files for man pages
to be sent to the remote repository.
2. Add reference to man page files in the new_docs/conf.py
and setup.py.
3. Add man pages that were manually built.
4. Change MANIFEST.in, removing old docs and adding new
docs sample config files.

References: BAR-404

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>

Add man page files to the remote repository

When building packages, the foundation packaging workflow
uses these files in the deb/debian/rules and
rpm/barman.spectemplate (and rpm/barman.foundationspectemplate)
to properly set the manuals for the commands by packages
(barman, barman-cli and barman-cli-cloud). These directories
mentioned previously are in the barman-packaging repository.

References: BAR-404

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>

Move files from /doc to /new_docs

These files are used as doc files for rpm and deb packaging.

They are present in the deb/debian/barman.install and in the
rpm/barman.spectemplate (and rpm/barman.foundationspectemplate).

They are required to properly build packages for the barman
software.

References: BAR-404

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>

rebase

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>