Refactor `_backup_report` method to deal with keep standalone targets
This commit refactors the _backup_report method for both window and
redundancy retention policies.

The code now handles the case when retention policy status of the
root backup is KEEP_STANDALONE and the incrementals are in or out of
the retention policy rule. For KEEP_FULL, it remains the same logic.
For KEEP_STANDALONE, when the root backup is within the retention policy
rule, the incrementals are considered valid. When this root backup falls
out of the retention policy rule, it is considered obsolete and will be
removed by the maintenance cron command.

This commit also modifies the `_propagate_retention_status_to_children`
so it does not have any logic regarding the retention policy. Now it just
propagates the status without checking the keep flag. This logic was moved
to the _backup_report method, so we handle FULL and STANDALONE KEEP flags.
The reason to move this was:
1. The checks for retention policy status should be a "business" logic for
_backup_report and the _propagate_retention_status_to_children should only
handle the propagation of the status.
2. We have both STANDALONE and FULL logic in the same place.

References: BAR-366

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>