Change config-switch argument description when using --help
This commit changes the model_name argument description from
the config-switch command to be clearer for a user that wants
to use the command. The model_name argument is an optional
argument but it's also a positional argument. As the help output
has positional argument and optional argument sections, the user
can get confused working with this command. The model_name
appears in the positional argument section of the help command.

References: BAR-378

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>