Change doc mentioning only MD5SUMS file, SHA256SUMS is included
This commit changes the documentation of the put-wal command that
was only mentioning the usage of MD5SUMS file for validation. In
BAR-155, we added support for FIPS compliance making sha256 the
default algorithm, so this needs to reflect in the whole doc.
Barman can still use md5, but it is not the default anymore.

This also adds the man pages where put-wal is mentioned.

References: BAR-437 & BAR-155

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>