Amend release note related with barman keep and incremental backups
In the release notes we have this message:

"`barman keep` command can only be applied to full backups when
`backup_method = postgres`. If a full backup has incremental backups
that depend on it, all of the incrementals are also kept by Barman."

In fact, the incremental backups are marked as valid not kept
(user can still delete them).

This commit amends the release note to be precise on this matter.

References: BAR-408

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>