Fix get_wal_conninfo and check_wal_streaming methods
This commit modifies the `check_wal_streaming` method in the Server
class and the get_wal_conninfo in the ServerConfig class.

Previously, when `conninfo` pointed to a standby server, Barman would
incorrectly use this connection to check the replication status,
resulting in the following error:

`replication slot (WAL streaming): FAILED (replication slot 'barman'
doesn't exist. Please execute 'barman receive-wal --create-slot pg17')`

The changes:

1. get_wal_conninfo()

When having wal-specific conninfos, wal_conninfo is not directly
dependent of wal_streaming_conninfo anymore. If wal_streaming_conninfo
is present and in the absence of wal_conninfo, wal_conninfo will fall back
to wal_streaming_conninfo. If wal_streaming_conninfo is not present,
wal_conninfo can still be set and used, and if it is not present, it will
fall back to conninfo.

2. check_wal_streaming

If wal_conninfo is different than conninfo, it means we have wal-specific
configuration and it must be use to check replication status. As this config
is not dependent of wal_streaming_conninfo, it can be used in the conditional.

References: BAR-409

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>