Refactor `walarchive` client to add FIPS support
This commit:

1. Adds `--md5` argument to the barman-wal-archive command.
	This is needed to keep compatibility between client
	and server version (PG server and Barman server). If
	the barman server is behind the PG server, this flag
	is needed so the client uses the `md5` hash algorithm
	to create the checksums. The barman server can only
	understand `md5` checksums at this point.

2. Changes  method name from `md5copyfileobj` to `hashCopyfileobj`
and refactor the code. A new paramter was added called `hash_algorithm`.
	This name is generic since we will use md5 or sha256
	depending on the case. The code was changed to accommodate
	the use of one of the two hash algorithms.

3. Adds an __init__ method to the ChecksumTarFile class so it is
initilized with hash_algorithm and HASHSUMS_FILE attributes for
`sha256`.
	This makes it easy to use the attributes in the class scope.
	The default value could have been set to `None` but it is
safer to keep a default value here since the methods using this
cannot receive None as an input.

4. Modifies the `addfile` method so it uses the new method
`hashCopyfileobj` with the new hash algorithm attribute as a parameter.

5. Modifies the `close` method so it uses generic names for variables.

6. Modifies the __init__ method from RemotePutWal class so it checks
for the new `--md5` flag, sets the correct algorithms and HASHSUMS file
and passes to the tar object.

7. Refactors `put_wal` command to keep compatibility between server and
client checksum calculations. The iteration on the tar file is in
streaming mode, so we cannot use random access. With this, we need to
iterate over the entire tar file, get the hash algorithm from the name
of the streamed HASHSUMS file and iterate over the dictionary to
calculate the checksums with the correct hash algorithm.

8. Refactored the `file_md5` method from utils so it has a generic name
to calculate with a hash algorithm. A new parameter was added
called `hash_algorithm`.

9. Add release_note for the feature - Add FIPS support

References: BAR-155

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>
Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>