Use `delete_this` annotation during the backup delete and cron processes
Add a delete annotation, which is applied to a backup at the
start of the delete process and only removed after it has finished.
The deletion process also ensures the `backup.info` file is the last
to be deleted. If a crash occurs, this file will remain, and the
system can resume the deletion upon recovery. In addition,
`cron_retention_policy` has been updated to check for backups marked
with this new annotation and, if found, the cron job will complete
the deletion process, preventing potential issues with orphaned files
and inaccurate backup size calculations.

References: BAR-244

Signed-off-by: Barbara Leidens <barbara.leidens@enterprisedb.com>
Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>
Co-authored-by: Martín Marqués <martin.marques@enterprisedb.com>
Co-authored-by: Israel <israel.barth@enterprisedb.com>