Modify release notes and include them in the docs
We removed the old `NEWS` file and replaced it with a new
`RELNOTES.md` file. The new file is better structured and
written in markdown language.

This new markdown file has been included in the new user
documentation as well, so they can access that info both
from GitHub and from the user docs.

References: BAR-290.

Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>
Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>