Add --keep-compression flag to barman-wal-restore and get-wal
A new `--keep-compression` option has been added to both `barman-wal-restore` and
`get-wal`. This option controls whether compressed WAL files should be decompressed
on the Barman server before being fetched. When specified with `get-wal`, default
decompression is skipped, and the output is the WAL file content in its original
state. When specified with `barman-wal-restore`, the WAL file is fetched as-is and,
if compressed, decompressed on the client side.

References: BAR-435

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>