Remove `sphinx` directory and rename `doc` as `old_docs`
The `sphinx` directory used to contain some very old scripts
used to build the docs through Sphinx in the past. This commit
removes that directory as it has not been used for a long time.

The `doc` directory contains the Barman documentation and man
pages that used to be used to build the docs up to Barman 3.11.1.
That used pandoc to achieve that goal. We have been refactoring
the docs of Barman, and the new version uses Sphinx to build
fresh new docs. With that in mind, we have renamed the `doc`
directory as `old_docs`, just so we can have a quick copy of any
of the old docs pages in the soon future, if needed. In any case,
that `old_docs` directory will be eventually removed from this
repository.

Other files were also updated in the repo so reflect these changes
in `sphinx` and `doc` directories:

* `linter.yml`: updated the regex to filter out `old_docs` instead of
  `doc` and `sphinx`.
* `INSTALL.md`: updated to simply redirect the user to the docs page.
* `README.rst`: temporarily updated renaming `doc` as `old_docs`. Also,
  removed the man pages on web, which no longer exist as a separate
  section.
* `recovery.rst`: updated the links to runbooks, to reference the ones
  from `old_docs`. Once we move them to the new docs, we will need to
  update the links again.

References: BAR-413.

Signed-off-by: Israel Barth <israel.barth@enterprisedb.com>