Add isoformat fields for backup start and end times in json output
This patch modifies the json output of the infofile object
adding two new fields: `begin_time_iso` and `end_time_iso`.
The new fields allow the use of a more standard and timezone aware
time format, preserving compatibility with previous versions.

Is worth noting that in the future the iso format for dates will be the
standard used by barman for storing dates and will be used everywhere
non human readable output is requested.

References: bar-494

Signed-off-by: Giulio Calacoci <giulio.calacoci@enterprisedb.com>