Add docs about running barman commands as postgres user
This adds documentation on how to setup sudo when running
barman on the postgres node.

We also added here subsubsections to recovery.rst for local
and remote recovery, by spliting the get-wal section into local
and remote recovery so it's clear that the sudo and /etc/sudoers
file changes are only for local recovery.

References: BAR-504

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>