Fix bug related to 'profile' variable not being referenced
An issue was introduced by BAR-242 as part of the Barman 3.12.0 release.

The issue was causing barman-cloud-backup-delete (and possibly other
commands) to fail with errors like this when --aws-profile argument or
aws_profile configuration were not set:

ERROR: Barman cloud backup delete exception: local variable 'profile'
referenced before assignment`

The problem occurred because the profile variable was unbound in a specific
code path: when config.aws_profile was defined, but with a falsy value (like
None). This typically happens when the config parameter is an instance of either
barman.config.ServerConfig or argparse.Namespace, and the user hasn't
configured aws_profile for Barman commands or didn't pass the --aws-profile
flag to barman-cloud-* scripts.
The issue is fixed by declaring the variable profile as None and assigning it to
the aws_profile config parameter if present.
Note: the original code, previous to BAR-242, used to set the profile variable
only when config.aws_region was defined. Although that was working, such dependency
between aws_profile and aws_region was not expected. The present commit still
addresses that concern, and at the same time also fixes the faulty logic
introduced by BAR-242.

References: BAR-518

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>
Co-authored-by: Israel Barth <israel.barth@enterprisedb.com>