Fix inconsistent diagnose output for retention policy info
This commit fixes the output of the `barman diagnose` command
by calling the __str__() method in the to_json() function,
specifically for the `retention_policy` and `wal_retention_policy`
fields. When the BarmanEncoderV2 is used to encode the json output,
it calls to_json() which will call the string representation of
the object. This change also impacts the output of the field
`wal_retention_policy`.

For the `retention_policy` and `wal_retention_policy` fields, the info
`REDUNDANCY 2` is more friendly to a user than `redundancy 2 b` and
`MAIN` is better than `simple-wal 2 b`, respectively.

References: BAR-100

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>