Fix issue with Barman not resolving NVMe EBS volumes in AWS
The implementation of `resolve_mounted_volume` is not 'resolving' NVMe volumes
in AWS when using Nitro instances.

This commit refactors the `resolve_mounted_volume` method from the class
`AwsVolumeMetadata` to fix this issue.

What changed?

The old implementation relied mainly on the AWS device name and a few
variations based on the virtualization_type. It uses '/dev/sd' as the base
and changes the prefix depending on whether the instance is HVM or
paravirtual.

The new implementation expands its identification strategy by considering
the '/sys/block' directory, which is part of the Linux sysfs virtual
filesystem. This approach looks for device names across different possible
block devices and also takes into account the device's serial number for
Nitro instances, improving robustness for finding the right device name.

References: BAR-500

Signed-off-by: Andre Marchesini <andre.marchesini@enterprisedb.com>