Improve Barman to Handle Duplicate WAL Files with Checksum Validation
This update ensures that Barman no longer rejects incoming WAL files
from Postgres archiving when a duplicate file is found in the
destination directory. Instead, Barman compares the checksums of the
incoming file and the existing file:

- If the checksums differ, an error message is logged, the incoming file
is moved to the to errors directory and the program will exit with 0.

- If the checksums are identical, a debug message is logged, the incoming
file is skipped and the program exits with 0.

This commit also creates a new method `move_duplicate_file_to_errors_directory`
to try to reuse this same logic in the Server's `put-wal` method and
in the WalArchiver's `archive` method. The archive command was slightly changed
to use this method in two places.

References: BAR-225

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>
Co-authored-by: Israel <israel.barth@enterprisedb.com>