Add and refactor unit tests for this patch
This commit changes the `test_put_wal_fail` so the cenario where
the incoming file already exists in the destination is tested apart
from this. It also adds the `test_put_wal_with_duplicate_file` to
tets the aforementioned cenario and the
`test_move_wal_file_to_errors_directory` to test the new
method added to the Server.

The test `test_archive` was also changed to accommodate the changes
from this PR.

References: BAR-225

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>