Extra: Add note about recommending single backup method per barman server
This commit adds a note recommending the use of a single backup method
per barman server. This was added here because there was a discussion
about possibly having snapshot backups mixed with other types of backups,
in the same catalog, which would require specific arguments to be passed
in the barman restore command. The user would not be able to know ahead
which backup would be used for restoration, so it would not be able to
know which parameters are required.

References: BAR-473

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>