Add utility function to parse target timeline
This commit adds a new utility function that parses the
target timeline based on the value passed by the user
when restoring a backup.

This function replaces two code blocks that were doing the
same thing and it is also used in the initial step of the
cli command `barman restore` to help fetch the correct
backup_id when the user wants to restore a backup without
a backup_id and recovery target timeline.

References: BAR-473

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>