Add support to restore without a backup_id
This commit adds support to restore a backup without providing
a backup_id in the command line.

The `restore` function of the cli.py was slightly refactored to
be able to handle a restoration without a backup id.
- No change when there is a backup_id passed.
- Recovery targets no allowed - `target_immediate`, `target_xid` and
`target_name`. There is an output error message.
- If no recovery_target is passed, the most recent backup is used
for the restore operation.
- If only `target_tli` is passed, the most recent backup in the
specified timeline is used.
- If `target_time` OR `target_lsn` is passed, the most recent backup
up to the specified `target_time` OR `target_lsn` is used.
- If `target_time` OR `target_lsn` is passed combined with a `target_tli`,
the most recent backup up to the specified `target_time` or `target_lsn`
in the specified timeline is used.
- If no backup_id is found, there is an output error message.

Methods to get the backup_id based on targets were added to keep with
the standard code structure of Barman.

References: BAR-473

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>