Add new xlogdb_directory config
Adds a new `xlogdb_directory` config option. It can be specified either
globally or per-server and it specifies a custom directory for the
xlogdb file.

The xlogdb file was also renamed. Regardless of whether the new
config is specified or not the new file name will be
`{servername}-xlog.db` instead of just `xlog.db`.

This change is a step forward in introducing ransomware protection in
Barman, as having the xlogdb file living in the barman server directory
stops it from being locked.

References: BAR-483

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>