Remove handling of staging WAL directory for Postgres < 8.4
This commit removes code that was handling a few things in regards to
the `barman_wal` directory on Postgres versions < 8.4. Such versions
do not have the `recovery_end_command` parameter and so Barman would
display a custom message at the end of a reovery with PITR reminding
the user to manually delete the `barman_wal` directory that was created.

This was done because a new option `--staging-wal-directory` is going to be added to
the restore/recover command and we do not want to deal with specific
cases caused by these older versions, which are already out of support.

References: BAR-224

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>