Remove WAL directories with rmtree when possible
When deleting WAL files, Barman iterates over them individually and
execute deletion one by one. This behavior has been reported to be
causing performance issues on ZFS filesystems.

In this sense, this commit adds an alternative way of deletion, deleting
the whole Barman WAL directory whenever possible.

The current code already had a list of WALs to be deleted and to be kept so
all I did was to introduce a new method `delete_wals`, which checks if every
WALs of a directory are in the removal list and, if so, delete the directory,
otherwise keep the current behavior of deleting each WAL individually.

When deleting a directory, in order to maintain the hook-scripts logic as it is
right now, 1) we run every pre-delete hook on every WAL; 2) remove the directory; 3) run
every post-delete hook on every WAL.

References: BAR-511

Signed-off-by: Gustavo William <gustavo.oliveira@enterprisedb.com>