Add support to barman-cloud-restore without a backup_id
This commit adds support to barman-cloud to restore a backup without
providing a backup_id in the command line.

What changed?
- The cloud_restore main function from clients.py was slightly refactored to
be able to handle a restoration without a backup id.
- `backup_id` was made optional.
- Added `target_time`, `target_lsn` and `target_tli` as command line options
(`target_lsn` and `target_time` are mutually exclusive.
- New method `get_latest_archived_wals_info` is added to the CloudBackupCatalog
class to be able to get the WalFileInfo from the last WAL file in the
archive. This is used to get the available timelines (in this case, the latest)
and it is used to fetch a backup when None is passed to recover.

How does it work?
- No change when there is a backup_id passed.
- Recovery targets allowed - `target_time`, `target_lsn` and `target_tli`.
- If no recovery_target is passed, the most recent backup is used
for the restore operation.
- If only `target_tli` is passed, the most recent backup in the
specified timeline is used.
- If `target_time` OR `target_lsn` is passed, the most recent backup
up to the specified `target_time` OR `target_lsn` is used.
- If `target_time` OR `target_lsn` is passed combined with a `target_tli`,
the most recent backup up to the specified `target_time` or `target_lsn`
in the specified timeline is used.
- If no backup_id is found, there is an output error message.

Docstrings were fixed for parse_target_tli because it can also receive a
CloudBackupCatalog object as a parameter.

References: BAR-541

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>