Allow PITR for any combination of recovery targets
This commit changes the method `set_pitr_targets` from the
RecoveryExecutor which assignes `True` to the value of the
`is_pitr` field of the recovery info dict.

Previously there were cases where Barman would not configure
Postgres for PITR, even if PITR was expected based on the
options specified for `barman restore`:

* If only `--target-tli` was given to `barman restore`; or
* If `--target-tli` value was the same as the backup timeline.

With this change, any combination of recovery targets, i.e. if
any `--target-*` option is passed to `barman restore`, that will
be treated as a PITR. That fixes the issues listed above when
only recovery target timeline is set or the timeline is the same
as the backup used for recovery.

References: BAR-543

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>