Add documentation about PITR
This commit adds information about PITR in the
`Point-in-Time Recovery` section.

What did we add?

- `--target-tli` option as recovery target.
- Info about the `recovery.signal` file creation.

The block with information about standby-mode was modified for
the sake of completeness.

Finally, two references to Postgres docs were added to give
directions to the user in case he/she needs more information
on archive recovery and recovery targets.

    Note: As a side change, we removed the info about Postgres10+
	  and --target-immediate being the default value. The first
	  info is irrelevant because the oldest supported version
	  is 13 and the second was just incorrect because Barman
	  default is to perform a simple backup recovery with
	  no PITR.

References: BAR-543

Signed-off-by: Andre <andre.marchesini@enterprisedb.com>
Co-authored-by: Israel Barth <israel.barth@enterprisedb.com>